AbstractGridArray constructors
