+++ src/RingGrids/general.jl
@@ -1,195 +1,183 @@
-abstract type AbstractGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractArray{T,N} end
-abstract type AbstractFullGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractGridArray{T,N} end
-abstract type AbstractReducedGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractGridArray{T,N} end
+abstract type AbstractGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractArray{T, N} end
+const AbstractGrid{T} = AbstractGridArray{T, 1, Vector{T}}
 
-# all AbstractGrids have their grid points stored in a vector field `data`
-# propagate length, size, getindex, setindex! for that
+# nonparametric_type(G::Type{<:AbstractGridArray}) = @warn "Please define nonparametric_type(::$G)"
+# horizontal_grid_type(G::Type{<:AbstractGridArray}) = @warn "Please define horizontal_grid_type(::$G)"
+
+# LENGTH SIZE
 Base.length(G::AbstractGridArray) = length(G.data)
 Base.size(G::AbstractGridArray) = size(G.data)
 Base.sizeof(G::AbstractGridArray) = sizeof(G.data)
 
-@inline Base.getindex(G::AbstractGridArray, ijk::Integer...) = getindex(G.data,ijk...)
-@inline Base.getindex(G::AbstractGridArray, r::AbstractRange) = get_index(G.data, r)
-
-
-
-@inline function Base.getindex(L::AbstractGridArray, col::Colon, I...)
-    return AbstractGridArray(getindex(L.data, col, I...), L.m, L.n)
-end
-
-
-
-@inline Base.getindex(L::AbstractGridArray, r::AbstractRange) = getindex(L.data,r)
-@inline Base.getindex(L::AbstractGridArray, r::AbstractRange, I...) = getindex(L.data, r, I...)
-@inline Base.getindex(L::AbstractGridArray, i::Integer) = getindex(L.data,i)
-@inline Base.getindex(L::AbstractGridArray, I::CartesianIndex) = getindex(L, Tuple(I)...)
-
-
-@inline Base.setindex!(G::AbstractGrid, x::AbstractVector, r::AbstractRange) = setindex!(G.data, x, r)
-@inline Base.setindex!(G::AbstractGrid, x, k::Integer) = setindex!(G.data, x, k)
-
-
-function grids_match(A::AbstractGridArray, B::AbstractGridArray)
-    length(A) == length(B) && return _grids_match(typeof(A), typeof(B))
-    return false
-end
-
-function _grids_match(A::Type{<:AbstractGridArray}, B::Type{<:AbstractGridArray})
-    return nonparametric_type(A) == nonparametric_type(B)
-end
-
-"""
-    abstract type AbstractGrid{T} <: AbstractVector{T} end
-
-The abstract supertype for all spatial grids on the sphere supported by SpeedyWeather.jl.
-Every new grid has to be of the form
-
-    abstract type AbstractGridClass{T} <: AbstractGrid{T} end
-    struct MyNewGrid{T} <: AbstractGridClass{T}
-        data::Vector{T}     # all grid points unravelled into a vector
-        nlat_half::Int      # resolution: latitude rings on one hemisphere (Equator incl)
-    end
-    
-`MyNewGrid` should belong to a grid class like `AbstractFullGrid`, `AbstractOctahedralGrid` or
-`AbstractHEALPixGrid` (that already exist but you may introduce a new class of grids) that share
-certain features such as the number of longitude points per latitude ring and indexing, but may
-have different latitudes or offset rotations. Each new grid `Grid` (or grid class) then has to
-implement the following methods (as an example, see octahedral.jl)
-    
-Fundamental grid properties
-    get_npoints         # total number of grid points
-    nlat_odd            # does the grid have an odd number of latitude rings?
-    get_nlat            # total number of latitude rings
-    get_nlat_half       # number of latitude rings on one hemisphere incl Equator
-    
-Indexing
-    get_nlon_max        # maximum number of longitudes points (at the Equator)
-    get_nlon_per_ring   # number of longitudes on ring j
-    each_index_in_ring  # a unit range that indexes all longitude points on a ring
-    
-Coordinates
-    get_colat           # vector of colatitudes (radians)
-    get_colatlon        # vectors of colatitudes, longitudes (both radians)
-
-Spectral truncation
-    truncation_order    # linear, quadratic, cubic = 1, 2, 3 for grid 
-    get_truncation      # spectral truncation given a grid resolution
-    get_resolution      # grid resolution given a spectral truncation
-
-Quadrature weights and solid angles
-    get_quadrature_weights  # = sinθ Δθ for grid points on ring j for meridional integration
-    get_solid_angle         # = sinθ Δθ Δϕ, solid angle of grid points on ring j
-"""
-AbstractGrid
-
-# Define methods that are universally applicable to any G<:AbstractGrid here
-# generator functions for grid
-Base.zeros(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid{T}} where T =
-    Grid(zeros(T, get_npoints(Grid, nlat_half)), nlat_half)
-# use Float64 if not provided
-Base.zeros(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid} = zeros(Grid{Float64}, nlat_half)
-# zero element of an AbstractGrid instance grid by packing a zero(::Vector) into grid
-Base.zero(grid::Grid) where {Grid<:AbstractGrid} = Grid(zero(grid.data))
-
-# initialise with ones
-Base.ones(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid{T}} where T =
-    Grid(ones(T, get_npoints(Grid, nlat_half)), nlat_half)
-Base.ones(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid} = ones(Grid{Float64}, nlat_half)
-
-# in case type parameter T in Grid{T} is provided
-function (::Type{Grid})(::UndefInitializer, nlat_half::Integer) where {Grid<:AbstractGrid{T}} where T
-    return Grid(Vector{T}(undef, get_npoints(Grid, nlat_half)), nlat_half)
-end
-
-# use Float64 if not
-function (::Type{Grid})(::UndefInitializer, nlat_half::Integer) where {Grid<:AbstractGrid}
-    return Grid(Vector{Float64}(undef, get_npoints(Grid, nlat_half)), nlat_half)
-end
-
-# randn initializer, use Float64 if T in Grid{T} not provided
-Base.randn(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid} = randn(Grid{Float64}, nlat_half)
-Base.randn(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid{T}} where T = 
-    Grid(randn(T, get_npoints(Grid, nlat_half)), nlat_half)
-
-# rand initializer, use Float64 if T in Grid{T} not provided
-Base.rand(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid} = rand(Grid{Float64}, nlat_half)
-Base.rand(::Type{Grid}, nlat_half::Integer) where {Grid<:AbstractGrid{T}} where T = 
-    Grid(rand(T, get_npoints(Grid, nlat_half)), nlat_half)
-
-# truncation is the spectral truncation corresponding to size of grid and lin/quad/cubic truncation
-get_resolution(grid::AbstractGrid) = get_nlat_half(grid)
-get_nlat_half(grid::AbstractGrid) = grid.nlat_half
-
-# does the grid have an odd number of latitudes?
-nlat_odd(grid::AbstractGrid) = nlat_odd(typeof(grid))
+get_nlat_half(grid::AbstractGridArray) = grid.nlat_half
+nlat_odd(grid::AbstractGridArray) = nlat_odd(typeof(grid))
 
 # get total number of latitude rings, *(nlat_half > 0) to return 0 for nlat_half = 0
-get_nlat(Grid::Type{<:AbstractGrid}, nlat_half::Integer) = 2nlat_half - nlat_odd(Grid)*(nlat_half > 0)
-get_nlat(grid::Grid) where {Grid<:AbstractGrid} = get_nlat(Grid, grid.nlat_half)
+get_nlat(Grid::Type{<:AbstractGridArray}, nlat_half::Integer) = 2nlat_half - nlat_odd(Grid)*(nlat_half > 0)
+get_nlat(grid::Grid) where {Grid<:AbstractGridArray} = get_nlat(Grid, grid.nlat_half)
 
-# get total number of grid pointst
-get_npoints(grid::Grid) where {Grid<:AbstractGrid} = get_npoints(Grid, grid.nlat_half)
+# get total number of grid points vs horizontal (2D) only
+get_npoints(grid::Grid) where {Grid<:AbstractGridArray} = get_npoints(Grid, grid.nlat_half)
+get_npoints(G::Type{<:AbstractGridArray}, nlat_half::Integer, k::Integer...) = prod(k) * get_npoints2D(G, nlat_half)
+get_npoints2D(grid::Grid) where {Grid<:AbstractGridArray} = get_npoints2D(Grid, grid.nlat_half)
 
-# coordinates
-get_latdlonds(grid::Grid) where {Grid<:AbstractGrid} = get_latdlonds(Grid, grid.nlat_half)
+## INDEXING
+@inline Base.getindex(G::AbstractGridArray, ijk::Integer...) = getindex(G.data,ijk...)
+@inline Base.getindex(G::AbstractGridArray, r::AbstractRange, k::Integer...) = getindex(G.data, r, k...)
+@inline Base.getindex(G::AbstractGridArray, ij::Integer, k::AbstractRange) = getindex(G.data, ij, k)
+@inline Base.getindex(G::AbstractGridArray, I::CartesianIndex) = getindex(G, Tuple(I)...)
 
-function get_latdlonds(Grid::Type{<:AbstractGrid}, nlat_half::Integer)
-    colats, lons = get_colatlons(Grid, nlat_half)     # colatitudes, longitudes in radians
+@inline function Base.getindex(
+    G::GridArray,
+    col::Colon,
+    k...,
+) where {GridArray<:AbstractGridArray{T, N, ArrayType}} where {T, N, ArrayType}
+    GridArray_ = nonparametric_type(GridArray)
+    return GridArray_{T, 1}(getindex(G.data, col, k...), G.nlat_half)
+end
+
+@inline Base.setindex!(G::AbstractGridArray, x, ijk::Integer...) =
+    setindex!(G.data, x, ijk...)
+@inline Base.setindex!(G::AbstractGridArray, x::AbstractVector, ij::AbstractRange, k::Integer...) =
+    setindex!(G.data, x, ij, k...)
+@inline Base.setindex!(G::AbstractGridArray, x::AbstractVector, ij::Integer, k::AbstractRange) =
+    setindex!(G.data, x, ij, k)
+
+## CONSTRUCTORS
+function check_inputs(data, nlat_half, rings, Grid)
+    check = true
+    check &= size(data, 1) == get_npoints2D(Grid, nlat_half)    # test number of 2D grid points
+    check &= length(rings) == get_nlat(Grid, nlat_half)         # test number of rings == nlat
+    # TODO also check that rings map to all and only valid grid points?
+    return check
+end
+
+function error_message(data, nlat_half, rings, G, T, N, A)
+    nlat = get_nlat(G, nlat_half)
+    nrings = length(rings)
+    if nlat != nrings
+        return error("$nrings-element ring indices "*
+            "cannot be used to create a $nlat-ring $G{$T, $N, $A}.")
+    else
+        return error("$(summary(data)) cannot be used to create a $nlat-ring $G{$T, $N, $A}")
+    end
+end
+
+# if no rings are provided, calculate them
+function (::Type{Grid})(data::AbstractArray, nlat_half::Integer) where {Grid<:AbstractGridArray}
+    GridArray_ = nonparametric_type(Grid)
+    rings = eachring(Grid, nlat_half)
+    return GridArray_(data, nlat_half, rings)
+end
+
+# if no nlat_half provided calculate it
+function (::Type{Grid})(data::AbstractArray) where {Grid<:AbstractGridArray}
+    npoints2D = size(data, 1)
+    nlat_half = get_nlat_half(Grid, npoints2D)
+    return Grid(data, nlat_half)
+end
+
+for f in (:zeros, :ones, :rand, :randn)
+    @eval begin
+        # general version with ArrayType(zeros(...)) conversion
+        function Base.$f(
+            ::Type{Grid},
+            nlat_half::Integer,
+            k::Integer...,
+        ) where {Grid<:AbstractGridArray{T, N, ArrayType}} where {T, N, ArrayType}
+            return Grid(ArrayType($f(T, get_npoints2D(Grid, nlat_half), k...)), nlat_half)
+        end
+
+        # CPU version with zeros(T, ...) producing Array
+        function Base.$f(
+            ::Type{Grid},
+            nlat_half::Integer,
+            k::Integer...,
+        ) where {Grid<:AbstractGridArray{T}} where T
+            return Grid($f(T, get_npoints2D(Grid, nlat_half), k...), nlat_half)
+        end
+
+        # use Float64 if no type provided
+        function Base.$f(
+            ::Type{Grid},
+            nlat_half::Integer,
+            k::Integer...,
+        ) where {Grid<:AbstractGridArray}
+            return $f(Grid{Float64}, nlat_half, k...)
+        end
+    end
+end
+
+# zero element of an AbstractGridArray instance grid by creating new zero(grid.data)
+Base.zero(grid::Grid) where {Grid<:AbstractGridArray} = Grid(zero(grid.data), grid.nlat_half, grid.rings)
+
+# general version with ArrayType{T, N}(undef, ...) generator
+function (::Type{Grid})(
+    ::UndefInitializer,
+    nlat_half::Integer,
+    k::Integer...,
+) where {Grid<:AbstractGridArray{T, N, ArrayType}} where {T, N, ArrayType}
+    return Grid(ArrayType{T, N}(undef, get_npoints2D(Grid, nlat_half), k...), nlat_half)
+end
+
+# CPU version with Array{T, N}(undef, ...) generator
+function (::Type{Grid})(
+    ::UndefInitializer,
+    nlat_half::Integer,
+    k::Integer...,
+) where {Grid<:AbstractGridArray{T}} where T
+    return Grid(Array{T}(undef, get_npoints2D(Grid, nlat_half), k...), nlat_half)
+end
+
+# use Float64 if no type provided
+function (::Type{Grid})(
+    ::UndefInitializer,
+    nlat_half::Integer,
+    k::Integer...,
+) where {Grid<:AbstractGridArray}
+    return Grid(Array{Float64}(undef, get_npoints2D(Grid, nlat_half), k...), nlat_half)
+end
+
+## COORDINATES
+get_latdlonds(grid::Grid) where {Grid<:AbstractGridArray} = get_latdlonds(Grid, grid.nlat_half)
+
+function get_latdlonds(Grid::Type{<:AbstractGridArray}, nlat_half::Integer)
+    colats, lons = get_colatlons(Grid, nlat_half)   # colatitudes, longitudes in radians
     latds = colats                                  # flat copy rename before conversion
     londs = lons
     latds .= π/2 .- colats                          # colatiudes to latitudes in radians
-    latds .= latds .* (360/2π)                      # now in degrees 90˚...-90˚
+    latds .*= (360/2π)                              # now in degrees 90˚...-90˚
     londs .*= (360/2π)
-
     return latds, londs
 end
 
-function get_latlons(Grid::Type{<:AbstractGrid}, nlat_half::Integer)
-    colats, lons = get_colatlons(Grid, nlat_half)     # colatitudes, longitudes in radians
+function get_latlons(Grid::Type{<:AbstractGridArray}, nlat_half::Integer)
+    colats, lons = get_colatlons(Grid, nlat_half)   # colatitudes, longitudes in radians
     lats = colats                                   # flat copy rename before conversion
     lats .= π/2 .- colats                           # colatiudes to latitudes in radians
-
     return lats, lons
 end
 
-get_lat(grid::Grid) where {Grid<:AbstractGrid} = get_lat(Grid, grid.nlat_half)
-get_latd(grid::Grid) where {Grid<:AbstractGrid} = get_latd(Grid, grid.nlat_half)
-get_lond(grid::Grid) where {Grid<:AbstractGrid} = get_lond(Grid, grid.nlat_half)
-get_lon(grid::Grid) where {Grid<:AbstractGrid} = get_lon(Grid, grid.nlat_half)
-get_colat(grid::Grid) where {Grid<:AbstractGrid} = get_colat(Grid, grid.nlat_half)
-get_colatlons(grid::Grid) where {Grid<:AbstractGrid} = get_colatlons(Grid, grid.nlat_half)
+get_lat(grid::Grid) where {Grid<:AbstractGridArray} = get_lat(Grid, grid.nlat_half)
+get_latd(grid::Grid) where {Grid<:AbstractGridArray} = get_latd(Grid, grid.nlat_half)
+get_lond(grid::Grid) where {Grid<:AbstractGridArray} = get_lond(Grid, grid.nlat_half)
+get_lon(grid::Grid) where {Grid<:AbstractGridArray} = get_lon(Grid, grid.nlat_half)
+get_colat(grid::Grid) where {Grid<:AbstractGridArray} = get_colat(Grid, grid.nlat_half)
+get_colatlons(grid::Grid) where {Grid<:AbstractGridArray} = get_colatlons(Grid, grid.nlat_half)
 
-function get_lat(Grid::Type{<:AbstractGrid}, nlat_half::Integer)
+function get_lat(Grid::Type{<:AbstractGridArray}, nlat_half::Integer)
     return π/2 .- get_colat(Grid, nlat_half)
 end
 
-function get_latd(Grid::Type{<:AbstractGrid}, nlat_half::Integer)
+function get_latd(Grid::Type{<:AbstractGridArray}, nlat_half::Integer)
     return get_lat(Grid, nlat_half) * (360/2π)
 end
 
 # only defined for full grids, empty vector as fallback
-get_lon(::Type{<:AbstractGrid}, nlat_half::Integer) = Float64[]
-get_lond(::Type{<:AbstractGrid}, nlat_half::Integer) = Float64[]
+get_lon(::Type{<:AbstractGridArray}, nlat_half::Integer) = Float64[]
+get_lond(::Type{<:AbstractGridArray}, nlat_half::Integer) = Float64[]
 
-"""
-    i = each_index_in_ring(grid, j)
-
-UnitRange `i` to access data on grid `grid` on ring `j`."""
-function each_index_in_ring(grid::Grid, j::Integer) where {Grid<:AbstractGrid}
-    return each_index_in_ring(Grid, j, grid.nlat_half)
-end
-
-"""
-    ijs = eachgridpoint(grid)
-
-UnitRange `ijs` to access each grid point on grid `grid`."""
-eachgridpoint(grid::AbstractGrid) = Base.OneTo(get_npoints(grid))
-function eachgridpoint(grid1::Grid, grids::Grid...) where {Grid<:AbstractGrid}
-    n = length(grid1)
-    Base._all_match_first(X->length(X), n, grid1, grids...) || throw(BoundsError)
-    return eachgridpoint(grid1)
-end
-
+## ITERATORS
 """
 $(TYPEDSIGNATURES)
 Vector{UnitRange} `rings` to loop over every ring of grid `grid`
@@ -199,9 +187,9 @@ and then each grid point per ring. To be used like
     for ring in rings
         for ij in ring
             grid[ij]"""
-eachring(grid::AbstractGrid) = eachring(typeof(grid), grid.nlat_half)
+eachring(grid::AbstractGridArray) = grid.rings
 
-function eachring(Grid::Type{<:AbstractGrid}, nlat_half::Integer)
+function eachring(Grid::Type{<:AbstractGridArray}, nlat_half::Integer)
     rings = Vector{UnitRange{Int}}(undef, get_nlat(Grid, nlat_half))
     each_index_in_ring!(rings, Grid, nlat_half)
     return rings
@@ -211,11 +199,37 @@ end
 $(TYPEDSIGNATURES)
 Same as `eachring(grid)` but performs a bounds check to assess that all grids
 in `grids` are of same size."""
-function eachring(grid1::Grid, grids::Grid...) where {Grid<:AbstractGrid}
+function eachring(grid1::Grid, grids::Grid...) where {Grid<:AbstractGridArray}
     @inline 
     n = length(grid1)
     Base._all_match_first(X->length(X), n, grid1, grids...) || throw(BoundsError)
     return eachring(grid1)
+end
+
+function grids_match(A::AbstractGridArray, B::AbstractGridArray)
+    length(A) == length(B) && return grids_match(typeof(A), typeof(B))
+    return false
+end
+
+function grids_match(A::Type{<:AbstractGridArray}, B::Type{<:AbstractGridArray})
+    return nonparametric_type(A) == nonparametric_type(B)
+end
+
+"""
+$(TYPEDSIGNATURES)
+UnitRange to access data on grid `grid` on ring `j`."""
+function each_index_in_ring(grid::Grid, j::Integer) where {Grid<:AbstractGridArray}
+    return each_index_in_ring(Grid, j, grid.nlat_half)
+end
+
+"""
+$(TYPEDSIGNATURES)
+UnitRange to access each grid point on grid `grid`."""
+eachgridpoint(grid::AbstractGridArray) = Base.OneTo(get_npoints(grid))
+function eachgridpoint(grid1::Grid, grids::Grid...) where {Grid<:AbstractGridArray}
+    n = length(grid1)
+    Base._all_match_first(X->length(X), n, grid1, grids...) || throw(BoundsError)
+    return eachgridpoint(grid1)
 end
 
 """
@@ -236,31 +250,34 @@ $(TYPEDSIGNATURES)
 Returns a vector `nlons` for the number of longitude points per latitude ring, north to south.
 Provide grid `Grid` and its resolution parameter `nlat_half`. For both_hemisphere==false only
 the northern hemisphere (incl Equator) is returned."""
-function get_nlons(Grid::Type{<:AbstractGrid}, nlat_half::Integer; both_hemispheres::Bool=false)
+function get_nlons(Grid::Type{<:AbstractGridArray}, nlat_half::Integer; both_hemispheres::Bool=false)
     n = both_hemispheres ? get_nlat(Grid, nlat_half) : nlat_half
     return [get_nlon_per_ring(Grid, nlat_half, j) for j in 1:n]
 end
 
-get_nlon_max(grid::Grid) where {Grid<:AbstractGrid} = get_nlon_max(Grid, grid.nlat_half)
+get_nlon_max(grid::Grid) where {Grid<:AbstractGridArray} = get_nlon_max(Grid, grid.nlat_half)
 
 ## BROADCASTING
-# following https://docs.julialang.org/en/v1/manual/interfaces/#man-interfaces-broadcasting
-import Base.Broadcast: BroadcastStyle, Broadcasted
 
-# {1} as grids are <:AbstractVector, Grid here is the non-parameteric Grid type!
-struct AbstractGridStyle{Grid} <: Broadcast.AbstractArrayStyle{1} end
+# # following https://docs.julialang.org/en/v1/manual/interfaces/#man-interfaces-broadcasting
+# import Base.Broadcast: BroadcastStyle, Broadcasted
 
-# important to remove Grid{T} parameter T (eltype/number format) here to broadcast
-# automatically across the same grid type but with different T
-# e.g. FullGaussianGrid{Float32} and FullGaussianGrid{Float64}
-Base.BroadcastStyle(::Type{Grid}) where {Grid<:AbstractGrid} =
-    AbstractGridStyle{nonparametric_type(Grid)}()
+# # {1} as grids are <:AbstractVector, Grid here is the non-parameteric Grid type!
+# struct AbstractGridArrayStyle{Grid} <: Broadcast.AbstractArrayStyle{1} end
 
-# allocation for broadcasting, create a new Grid with undef of type/number format T
-function Base.similar(bc::Broadcasted{AbstractGridStyle{Grid}}, ::Type{T}) where {Grid, T}
-    Grid(Vector{T}(undef,length(bc)))
-end
+# # important to remove Grid{T} parameter T (eltype/number format) here to broadcast
+# # automatically across the same grid type but with different T
+# # e.g. FullGaussianGrid{Float32} and FullGaussianGrid{Float64}
+# Base.BroadcastStyle(::Type{Grid}) where {Grid<:AbstractGridArray} =
+#     AbstractGridArrayStyle{nonparametric_type(Grid)}()
 
-# ::Val{0} for broadcasting with 0-dimensional, ::Val{1} for broadcasting with vectors
-AbstractGridStyle{Grid}(::Val{0}) where Grid = AbstractGridStyle{Grid}()
-AbstractGridStyle{Grid}(::Val{1}) where Grid = AbstractGridStyle{Grid}()
+# # allocation for broadcasting, create a new Grid with undef of type/number format T
+# function Base.similar(bc::Broadcasted{AbstractGridArrayStyle{Grid}}, ::Type{T}) where {Grid, T}
+#     Grid(Vector{T}(undef,length(bc)))
+# end
+
+# # ::Val{0} for broadcasting with 0-dimensional, ::Val{1} for broadcasting with vectors, etc
+# AbstractGridArrayStyle{Grid}(::Val{0}) where Grid = AbstractGridArrayStyle{Grid}()
+# AbstractGridArrayStyle{Grid}(::Val{1}) where Grid = AbstractGridArrayStyle{Grid}()
+# AbstractGridArrayStyle{Grid}(::Val{2}) where Grid = AbstractGridArrayStyle{Grid}()
+# AbstractGridArrayStyle{Grid}(::Val{3}) where Grid = AbstractGridArrayStyle{Grid}()
