+++ src/RingGrids/grids/octahedral_clenshaw.jl
